requirements.txt
open-clip-torch==2.26.1
faiss-cpu
torchvision
fastapi
uvicorn
python-multipart
pytesseract
Pillow
numpy
rapidfuzz
requests
psycopg2-binary
minio
PyJWT
bcrypt
